refactor: consolidate label printing and batch management logic into dedicated stock modules while removing obsolete doctypes
