+++ erpnextkta/api.py
@@ -14,6 +14,20 @@ from erpnextkta.kta_sales.doctype.kta_sales_order_update_comparison.kta_sales_or
     compare_sales_order_update_documents as _compare_sales_order_update_documents,
 )
 
+
+# Refactored Imports
+from erpnextkta.kta_stock.label_manager import (
+    print_kta_pr_labels,
+    print_split_kta_pr_labels,
+    print_kta_wo_labels,
+    print_kta_wo_labels_of_stock_entry,
+    resplit_and_print_kta_wo_labels
+)
+from erpnextkta.kta_stock.batch_manager import (
+    BatchSplitManager,
+    check_packaging_quantity_mismatch
+)
+
 # Global doctype constants
 DOCTYPE_PARTY_ACCOUNT = "Party Account"
 DOCTYPE_CUSTOMER = "Customer"
@@ -133,257 +147,20 @@ def get_customer_income_account(customer, company):
         return None
 
 
-@frappe.whitelist()
-def print_kta_pr_labels(gr_number=None, label=None, q_ref=None):
-    if not gr_number and not label and not q_ref:
-        frappe.msgprint("Either `gr_number`, `label` or 'q_ref' must be provided.")
-        return
-
-    query_filter = {FIELD_DO_NOT_SPLIT: 0}
-    if gr_number:
-        query_filter[FIELD_GR_NUMBER] = gr_number
-    elif label:
-        query_filter[FIELD_NAME] = label
-    elif q_ref:
-        query_filter[FIELD_QUALITY_REF] = q_ref
-
-    zebra_printer = get_zebra_printer_for_user()
-    zebra_ip_address = zebra_printer.get("ip")
-    zebra_port = zebra_printer.get("port")
-
-    for data in frappe.get_all(
-            doctype=DOCTYPE_KTA_DEPO_ETIKETLERI,
-            filters=query_filter,
-            fields={
-                FIELD_ITEM_CODE,
-                FIELD_ITEM_NAME,
-                FIELD_ITEM_GROUP,
-                FIELD_QTY,
-                FIELD_UOM,
-                FIELD_SUPPLIER_DELIVERY_NOTE,
-                FIELD_SUT_BARCODE,
-                FIELD_GR_POSTING_DATE,
-                FIELD_QUALITY_REF
-            }
-    ):
-        data.qty = format_kta_label_qty(data.qty)
-        formatted_data = zebra_formatter(DOCTYPE_KTA_DEPO_ETIKETLERI, data)
-        send_data_to_zebra(formatted_data, zebra_ip_address, zebra_port)
 
 
-@frappe.whitelist()
-def print_split_kta_pr_labels(label=None):
-    if not label:
-        frappe.msgprint("`label` must be provided.")
-        return
-
-    split_query_filter = {FIELD_PARENT: label}
-
-    splits = frappe.get_all(
-        doctype=DOCTYPE_KTA_DEPO_ETIKETLERI_BOLME,
-        filters=split_query_filter,
-        fields={FIELD_IDX, FIELD_QTY}
-    )
-
-    query_filter = {FIELD_DO_NOT_SPLIT: 1, FIELD_NAME: label}
-
-    label_data = frappe.db.get_value(
-        doctype=DOCTYPE_KTA_DEPO_ETIKETLERI,
-        filters=query_filter,
-        fieldname=[
-            FIELD_ITEM_CODE,
-            FIELD_ITEM_NAME,
-            FIELD_ITEM_GROUP,
-            FIELD_QTY,
-            FIELD_UOM,
-            FIELD_SUPPLIER_DELIVERY_NOTE,
-            FIELD_BATCH,
-            FIELD_SUT_BARCODE,
-            FIELD_GR_POSTING_DATE,
-            FIELD_QUALITY_REF
-        ],
-        as_dict=True
-    )
-
-    zebra_printer = get_zebra_printer_for_user()
-    zebra_ip_address = zebra_printer.get("ip")
-    zebra_port = zebra_printer.get("port")
-
-    base_batch = label_data.batch[:7] if label_data.batch and len(label_data.batch) > 7 else label_data.batch
-    for split in splits:
-        label_data.qty = format_kta_label_qty(split.qty)
-        label_data.batch = base_batch  # batch alanı ana bundle ID'si olarak sabitlenir
-        label_data.sut_barcode = f"{base_batch}{split.idx:04d}"
-        formatted_data = zebra_formatter(DOCTYPE_KTA_DEPO_ETIKETLERI, label_data)
-        send_data_to_zebra(formatted_data, zebra_ip_address, zebra_port)
 
 
-@frappe.whitelist()
-def print_kta_wo_labels(work_order):
-    details_of_wo = get_details_of_wo_for_label(work_order)
-
-    for stock_entry in frappe.get_all(
-            doctype=DOCTYPE_STOCK_ENTRY,
-            filters={
-                FIELD_STOCK_ENTRY_TYPE: VALUE_MANUFACTURE,
-                FIELD_WORK_ORDER: work_order
-            },
-            fields=[FIELD_NAME]
-    ):
-        print_kta_wo_label(details_of_wo, stock_entry.name)
 
 
-@frappe.whitelist()
-def print_kta_wo_labels_of_stock_entry(stock_entry):
-    stock_entry_doc = frappe.get_doc(DOCTYPE_STOCK_ENTRY, stock_entry)
-    print_kta_wo_label(get_details_of_wo_for_label(stock_entry_doc.get(FIELD_WORK_ORDER)), stock_entry)
 
 
-def get_details_of_wo_for_label(work_order):
-    work_order_doc = frappe.get_doc(DOCTYPE_WORK_ORDER, work_order)
-    bom_doc = frappe.get_doc(DOCTYPE_BOM, work_order_doc.get(FIELD_BOM_NO))
-
-    material_index = "-"
-    meta = frappe.get_meta(DOCTYPE_BOM)
-    if meta.has_field(FIELD_CUSTOM_MUSTERI_INDEKSI_NO):
-        material_index = bom_doc.get(FIELD_CUSTOM_MUSTERI_INDEKSI_NO)
-
-    musteri_paketleme_miktari = frappe.db.get_value(
-        doctype=DOCTYPE_ITEM_CUSTOMER_DETAIL,
-        filters={
-            FIELD_PARENT: work_order_doc.get(FIELD_PRODUCTION_ITEM),
-            FIELD_PARENTTYPE: DOCTYPE_ITEM,
-            FIELD_PARENTFIELD: VALUE_CUSTOMER_ITEMS
-        },
-        fieldname=[f"max({FIELD_CUSTOM_MUSTERI_PAKETLEME_MIKTARI}) as musteri_paketleme_miktari"]
-    )
-
-    if not musteri_paketleme_miktari:
-        frappe.throw(
-            f"No {FIELD_CUSTOM_MUSTERI_PAKETLEME_MIKTARI} found for Item: {work_order_doc.get(FIELD_PRODUCTION_ITEM)}")
-        return None
-
-    return {
-        FIELD_WORK_ORDER: work_order_doc.get(FIELD_NAME),
-        FIELD_DESCRIPTION: work_order_doc.get(FIELD_DESCRIPTION),
-        FIELD_STOCK_UOM: work_order_doc.get(FIELD_STOCK_UOM),
-        FIELD_PRODUCTION_ITEM: work_order_doc.get(FIELD_PRODUCTION_ITEM),
-        "material_index": material_index,
-        "musteri_paketleme_miktari": musteri_paketleme_miktari
-    }
 
 
-def print_kta_wo_label(work_order_details, stock_entry):
-    source_warehouse = stock_entry
-
-    stock_entry_detail = frappe.get_all(
-        doctype=DOCTYPE_STOCK_ENTRY_DETAIL,
-        filters={
-            FIELD_PARENT: stock_entry,
-            FIELD_PARENTTYPE: DOCTYPE_STOCK_ENTRY,
-            FIELD_PARENTFIELD: PARENT_FIELD_STOCK_ENTRY_DETAIL,
-            FIELD_ITEM_CODE: work_order_details.get(FIELD_PRODUCTION_ITEM),
-            FIELD_IS_FINISHED_ITEM: 1,
-            FIELD_DOCSTATUS: 1,
-            FIELD_T_WAREHOUSE: ["is", "set"]
-        },
-        fields=[FIELD_NAME],
-        as_list=True
-    )
-
-    if len(stock_entry_detail) > 1:
-        frappe.throw(f"More than one Inward Type of Transaction found for Stock Entry: {stock_entry}")
-        return
-
-    stock_entry_detail_doc = frappe.get_doc(DOCTYPE_STOCK_ENTRY_DETAIL, stock_entry_detail[0])
-    stock_entry_doc = frappe.get_doc(DOCTYPE_STOCK_ENTRY, stock_entry)
-
-    destination_warehouse = stock_entry_doc.get(FIELD_TO_WAREHOUSE)
-    if not destination_warehouse:
-        destination_warehouse = stock_entry_detail_doc.get(FIELD_T_WAREHOUSE)
-
-    batch_no = get_batch_from_stock_entry_detail(stock_entry_detail_doc)
-    base_batch_no = get_base_batch_from_work_order(work_order_details.get(FIELD_WORK_ORDER)) or batch_no
-
-    # Construct data
-    data = frappe.get_doc({
-        'doctype': DOCTYPE_KTA_IS_EMRI_ETIKETLERI,
-        'print_date': frappe.utils.nowdate(),
-        'material_number': work_order_details.get(FIELD_PRODUCTION_ITEM),
-        'material_description': work_order_details.get(FIELD_DESCRIPTION),
-        'material_index': work_order_details.get("material_index"),
-        FIELD_WORK_ORDER: work_order_details.get(FIELD_WORK_ORDER),
-        FIELD_GR_POSTING_DATE: frappe.utils.get_date_str(stock_entry_doc.get(FIELD_POSTING_DATE)),
-        FIELD_GR_NUMBER: stock_entry,
-        'gr_source_warehouse': source_warehouse,
-        FIELD_TO_WAREHOUSE: destination_warehouse,
-        FIELD_STOCK_UOM: work_order_details.get(FIELD_STOCK_UOM),
-        FIELD_BATCH_NO: base_batch_no
-    })
-
-    zebra_printer = get_zebra_printer_for_user()
-    zebra_ip_address = zebra_printer.get("ip")
-    zebra_port = zebra_printer.get("port")
-
-    batch_entries = []
-    if stock_entry_detail_doc.get("serial_and_batch_bundle"):
-        batch_entries = frappe.get_all(
-            DOCTYPE_SERIAL_AND_BATCH_ENTRY,
-            filters={
-                FIELD_PARENT: stock_entry_detail_doc.get("serial_and_batch_bundle"),
-                FIELD_PARENTTYPE: DOCTYPE_SERIAL_AND_BATCH_BUNDLE,
-                FIELD_PARENTFIELD: VALUE_ENTRIES,
-                FIELD_IS_OUTWARD: 0,
-                FIELD_DOCSTATUS: 1,
-                FIELD_BATCH_NO: ["is", "set"],
-            },
-            fields=[FIELD_BATCH_NO, FIELD_QTY],
-            order_by="idx asc",
-        )
-
-    if batch_entries:
-        for entry in batch_entries:
-            data.qty = format_kta_label_qty(entry.get(FIELD_QTY))
-            data.batch_no = base_batch_no
-            data.sut_no = entry.get(FIELD_BATCH_NO)
-            formatted_data = zebra_formatter(DOCTYPE_KTA_IS_EMRI_ETIKETLERI, data)
-            send_data_to_zebra(formatted_data, zebra_ip_address, zebra_port)
-    else:
-        musteri_paketleme_miktari = work_order_details.get("musteri_paketleme_miktari")
-        num_packs = frappe.cint(stock_entry_detail_doc.get(FIELD_QTY) // musteri_paketleme_miktari)
-        remainder_qty = stock_entry_detail_doc.get(FIELD_QTY) % musteri_paketleme_miktari
-
-        if num_packs >= 1:
-            for pack in range(1, num_packs + 1):
-                data.qty = format_kta_label_qty(musteri_paketleme_miktari)
-                data.sut_no = f"{batch_no}{pack:04d}"
-                formatted_data = zebra_formatter(DOCTYPE_KTA_IS_EMRI_ETIKETLERI, data)
-                send_data_to_zebra(formatted_data, zebra_ip_address, zebra_port)
-
-        if remainder_qty > 0:
-            data.qty = format_kta_label_qty(remainder_qty)
-            data.sut_no = f"{batch_no}{num_packs + 1:04d}"
-            formatted_data = zebra_formatter(DOCTYPE_KTA_IS_EMRI_ETIKETLERI, data)
-            send_data_to_zebra(formatted_data, zebra_ip_address, zebra_port)
-
-    data.delete()
 
 
-def send_data_to_zebra(data, ip, port):
-    try:
-        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-            s.settimeout(5)
-            s.connect((ip, port))
-            s.sendall(data.encode("utf-8"))
-            return None
-    except Exception as e:
-        frappe.log_error(title=f"ZPL Print Error {str(e)}", message="Printer Error")
-        return {"status": "error", "message": f"Failed to send label {str(e)}"}
 
 
-def zebra_formatter(doctype_name, data):
-    doc = frappe.get_doc(DOCTYPE_KTA_ZEBRA_TEMPLATES, doctype_name)
-    return doc.get("zebra_template").format(data=data)
 
 
 def custom_split_kta_batches(row=None, q_ref="ATLA 5/1"):
@@ -451,155 +228,6 @@ def custom_split_kta_batches(row=None, q_ref="ATLA 5/1"):
             q_ref=q_ref,
         )
 
-def _get_reliable_incoming_rate(row, bundle_doc):
-    """
-    Güvenilir incoming_rate belirler.
-    Fallback zinciri (öncelik sırasıyla):
-    0. Submitted bundle ise bağlı SLE'nin incoming_rate'i (QI-path için kritik)
-    1. Bundle entry'lerindeki mevcut incoming_rate
-    2. DB'den PR Item / SE Detail valuation_rate (taze okuma)
-    3. DB'den PR Item rate veya base_rate
-    4. Item Master valuation_rate
-    5. Item Master last_purchase_rate
-
-    Öncelik 0'ın önemi:
-    QI (Kalite Kontrol) akışında bundle, SLE oluşturulduktan SONRA değiştiriliyor.
-    Eski bundle entry'leri incoming_rate=0 içerebilir ya da farklı bir rate döndürebilir.
-    SLE zaten doğru rate ile kaydedilmiş, bu yüzden oradan okumak en güvenli.
-    """
-    # 0. Submitted bundle için bağlı SLE'den incoming_rate al
-    # Bu, QI onayı sırasında split edilen bundle'lar için kritik:
-    # SLE zaten doğru rate ile oluşturulmuş olup bundle entry'lerinden daha güvenilir.
-    if bundle_doc.name and flt(bundle_doc.docstatus) == 1:
-        sle_rate = frappe.db.get_value(
-            "Stock Ledger Entry",
-            {
-                "serial_and_batch_bundle": bundle_doc.name,
-                "actual_qty": (">", 0),
-                "is_cancelled": 0,
-                "docstatus": 1,
-            },
-            "incoming_rate",
-        )
-        if flt(sle_rate):
-            return flt(sle_rate)
-
-    # 1. Bundle entry'lerinden
-    for entry in bundle_doc.get("entries", []):
-        if flt(entry.get("incoming_rate")):
-            return flt(entry.incoming_rate)
-
-    # 2-3. DB'den PR Item veya SE Detail satırını taze oku
-    if row.doctype == "Purchase Receipt Item":
-        db_values = frappe.db.get_value(
-            "Purchase Receipt Item", row.name,
-            ["valuation_rate", "rate", "base_rate"],
-            as_dict=True
-        )
-        if db_values:
-            rate = flt(db_values.valuation_rate) or flt(db_values.rate) or flt(db_values.base_rate)
-            if rate:
-                return rate
-    elif row.doctype == "Stock Entry Detail":
-        db_values = frappe.db.get_value(
-            "Stock Entry Detail", row.name,
-            ["valuation_rate", "basic_rate"],
-            as_dict=True
-        )
-        if db_values:
-            rate = flt(db_values.valuation_rate) or flt(db_values.basic_rate)
-            if rate:
-                return rate
-
-    # In-memory row values (son çare olarak row objesinden)
-    rate = flt(row.get("valuation_rate")) or flt(row.get("rate"))
-    if rate:
-        return rate
-
-    # 4. Item Master valuation_rate
-    item_valuation = frappe.db.get_value("Item", row.item_code, "valuation_rate")
-    if flt(item_valuation):
-        return flt(item_valuation)
-
-    # 5. Son satınalma oranı (Last Purchase Rate)
-    last_purchase_rate = frappe.db.get_value("Item", row.item_code, "last_purchase_rate")
-    if flt(last_purchase_rate):
-        return flt(last_purchase_rate)
-
-    frappe.log_error(
-        f"Incoming rate 0 olarak belirlendi! Row: {row.name}, Item: {row.item_code}, "
-        f"Bundle: {bundle_doc.name}",
-        "KTA Stock Valuation Warning"
-    )
-    return 0.0
-
-
-def _update_bundle_safely(row, allocations):
-    """
-    Serial and Batch Bundle'ı yeni allocation'lar ile günceller.
-    Optimized to minimize DB hits and avoid recursive validations.
-    """
-    bundle_name = row.get("serial_and_batch_bundle")
-    if not bundle_name:
-        return
-
-    bundle_doc = frappe.get_doc("Serial and Batch Bundle", bundle_name)
-    
-    bundle_doc.flags.ignore_validate = True
-    bundle_doc.flags.ignore_validate_update_after_submit = True
-    bundle_doc.flags.ignore_links = True
-    
-    warehouse = (
-        row.get(FIELD_WAREHOUSE)
-        or row.get(FIELD_T_WAREHOUSE)
-        or row.get(FIELD_S_WAREHOUSE)
-    )
-    
-    if not warehouse:
-        frappe.throw(_("Warehouse not found for row {0}").format(row.name))
-
-    original_incoming_rate = _get_reliable_incoming_rate(row, bundle_doc)
-
-    bundle_doc.set("entries", [])
-    total_qty = 0
-    total_amount = 0.0
-    for alloc in allocations:
-        alloc_qty = flt(alloc["qty"])
-        stock_value_diff = alloc_qty * original_incoming_rate
-        bundle_doc.append("entries", {
-            "batch_no": alloc["batch_no"],
-            "qty": alloc_qty,
-            "warehouse": warehouse,
-            "is_outward": 0,
-            "incoming_rate": original_incoming_rate,
-            "stock_value_difference": stock_value_diff,
-        })
-        total_qty += alloc_qty
-        total_amount += stock_value_diff
-    
-    bundle_doc.total_qty = total_qty
-    bundle_doc.total_amount = total_amount
-    if total_qty:
-        bundle_doc.avg_rate = total_amount / total_qty
-    
-    # Use save(ignore_permissions=True) to trigger a clean update with our flags
-    bundle_doc.save(ignore_permissions=True)
-    frappe.clear_document_cache("Serial and Batch Bundle", bundle_name)
-
-    # Batch qty güncelle: çocuk batch'lerin qty'sini set et, ana batch'i düşür
-    parent_batch_nos = set()
-    for alloc in allocations:
-        child_batch = alloc["batch_no"]
-        child_qty = flt(alloc["qty"])
-        # Çocuk batch'in qty'sini artır
-        current_qty = flt(frappe.db.get_value("Batch", child_batch, "batch_qty"))
-        frappe.db.set_value("Batch", child_batch, "batch_qty", current_qty + child_qty,
-                           update_modified=False)
-
-    # Ana batch'in qty'sini bundle entry'lerden yeniden hesapla
-    if row.get("batch_no"):
-        _recalculate_batch_qty(row.batch_no)
-
 
 
 
@@ -628,392 +256,22 @@ def custom_create_packages(row, batch_no, qty, sut_code, q_ref):
     etiket.insert(ignore_permissions=True)
 
 
-def _get_last_batch_number_for_base(base_batch_number):
-    if not base_batch_number: return 0
-
-    prefix_len = len(base_batch_number)
-    total_len = prefix_len + 4
-
-    # Sadece tam olarak [BASE][4 haneli rakam] formatındakileri bul
-    last_batch = frappe.db.sql("""
-        SELECT name FROM `tabBatch`
-        WHERE name LIKE %s
-        AND LENGTH(name) = %s
-        ORDER BY name DESC LIMIT 1
-    """, (f"{base_batch_number}%", total_len))
-
-    if not last_batch:
-        return 0
-
-    suffix = last_batch[0][0][prefix_len:]
-    return int(suffix) if suffix.isdigit() else 0
-
-
-def _prepare_batch_allocations(row, base_source_doc, base_batch_number, is_manufacturing=False):
-    """
-    Generic allocation logic for both PR and Manufacturing.
-    """
-    if is_manufacturing:
-        qty = flt(row.get("transfer_qty") or row.get("qty") or row.get("stock_qty") or 0)
-        split_qty = _get_customer_packaging_qty(row.item_code)
-    else:
-        qty = flt(row.stock_qty or 0)
-        split_qty = flt(row.custom_split_qty or 0)
-
-    if not qty or split_qty <= 0:
-        return []
-
-    if not is_manufacturing and row.get("custom_do_not_split"):
-        return [{
-            "batch_no": base_batch_number,
-            "qty": qty,
-            "sut_code": f"{base_batch_number}{0:04d}",
-            "pack_no": 0,
-        }]
-
-    num_packs = cint(qty // split_qty)
-    remainder_qty = qty % split_qty
-    allocations = []
-
-    # Get last sequence for manufacturing to allow continuous split numbering
-    start_pack_no = 1
-    if is_manufacturing:
-        start_pack_no = _get_last_batch_number_for_base(base_batch_number) + 1
-
-    for i in range(num_packs):
-        pack_no = start_pack_no + i
-        batch_no = _create_split_batch_record(row, base_source_doc, base_batch_number, pack_no, is_manufacturing)
-        allocations.append({
-            "batch_no": batch_no,
-            "qty": split_qty,
-            "sut_code": f"{base_batch_number}{pack_no:04d}",
-            "pack_no": pack_no,
-        })
-
-    if remainder_qty > 0:
-        pack_no = start_pack_no + num_packs
-        batch_no = _create_split_batch_record(row, base_source_doc, base_batch_number, pack_no, is_manufacturing)
-        allocations.append({
-            "batch_no": batch_no,
-            "qty": remainder_qty,
-            "sut_code": f"{base_batch_number}{pack_no:04d}",
-            "pack_no": pack_no,
-        })
-
-    return allocations
-
-
-def _recalculate_batch_qty(batch_no):
-    """Batch'in qty'sini Serial and Batch Bundle entry'lerden yeniden hesaplar."""
-    result = frappe.db.sql("""
-        SELECT COALESCE(SUM(sabe.qty), 0) as total_qty
-        FROM `tabSerial and Batch Entry` sabe
-        JOIN `tabSerial and Batch Bundle` sabb ON sabe.parent = sabb.name
-        WHERE sabe.batch_no = %s
-          AND sabb.is_cancelled = 0
-          AND sabb.docstatus = 1
-    """, batch_no)
-    new_qty = flt(result[0][0]) if result else 0.0
-    frappe.db.set_value("Batch", batch_no, "batch_qty", new_qty, update_modified=False)
-
-
-def _create_split_batch_record(row, parent_doc, base_batch_number, pack_no, is_manufacturing):
-    batch_id = f"{base_batch_number}{pack_no:04d}"
-    
-    if frappe.db.exists("Batch", batch_id):
-        return batch_id
-
-    batch_doc = frappe.get_doc({
-        "doctype": "Batch",
-        "batch_id": batch_id,
-        "item": row.item_code,
-        "stock_uom": row.get("stock_uom"),
-        "description": row.get("description"),
-        "batch_qty": 0,
-    })
-
-    if is_manufacturing:
-        batch_doc.update({
-            "reference_doctype": "Work Order",
-            "reference_name": parent_doc.work_order,
-            "stock_entry_reference_doctype": "Stock Entry",
-            "stock_entry_reference_name": parent_doc.name,
-            "manufacturing_date": parent_doc.get(FIELD_POSTING_DATE),
-        })
-    else:
-        batch_doc.update({
-            "supplier": parent_doc.get("supplier"),
-            "reference_doctype": "Purchase Receipt",
-            "reference_name": parent_doc.name,
-            "manufacturing_date": row.get("manufacturing_date") or parent_doc.get(FIELD_POSTING_DATE),
-            "expiry_date": row.get("expiry_date"),
-        })
-
-    batch_doc.flags.ignore_permissions = True
-    batch_doc.insert()
-    return batch_doc.name
-
-
-
-def _get_single_inward_batch_entry(bundle_name):
-    entries = frappe.get_all(
-        DOCTYPE_SERIAL_AND_BATCH_ENTRY,
-        filters={
-            FIELD_PARENT: bundle_name,
-            FIELD_PARENTTYPE: DOCTYPE_SERIAL_AND_BATCH_BUNDLE,
-            FIELD_IS_OUTWARD: 0,
-        },
-        fields=[FIELD_BATCH_NO, FIELD_QTY],
-        order_by="idx asc",
-    )
-
-    if len(entries) != 1:
-        return None
-
-    return entries[0]
-
-
-def _get_customer_packaging_qty(item_code):
-    result = frappe.get_all(
-        DOCTYPE_ITEM_CUSTOMER_DETAIL,
-        filters={
-            FIELD_PARENT: item_code,
-            FIELD_PARENTTYPE: DOCTYPE_ITEM,
-            FIELD_PARENTFIELD: VALUE_CUSTOMER_ITEMS,
-        },
-        fields=[f"max({FIELD_CUSTOM_MUSTERI_PAKETLEME_MIKTARI}) as packaging_qty"],
-        limit=1,
-    )
-
-    if not result:
-        return 0
-
-    return flt(result[0].packaging_qty or 0)
-
 
 # These are now merged into _prepare_batch_allocations and _create_split_batch_record
 
 
-def _create_manufacturing_split_batch(row, stock_entry, base_batch_number, pack_no):
-    """Yeni bir manufacturing batch oluşturur"""
-    batch_id = f"{base_batch_number}{pack_no:04d}"
-    
-    # Batch zaten varsa, mevcut olanı kullan
-    if frappe.db.exists("Batch", batch_id):
-        return batch_id
-
-    try:
-        batch_doc = frappe.get_doc({
-            "doctype": "Batch",
-            "batch_id": batch_id,
-            "item": row.item_code,
-            "reference_doctype": "Work Order",
-            "reference_name": stock_entry.work_order,
-            "stock_entry_reference_doctype": "Stock Entry",
-            "stock_entry_reference_name": stock_entry.name,
-        })
-        batch_doc.flags.ignore_permissions = True
-        batch_doc.insert()
-        return batch_doc.name
-    except Exception as e:
-        frappe.log_error(f"Batch Oluşturma Hatası: {str(e)}", "KTA Split Error")
-        raise e
-
-    # Work Order ile ilişkilendirme
-    work_order = None
-    if hasattr(stock_entry, 'work_order'):
-        work_order = getattr(stock_entry, 'work_order', None)
-    elif isinstance(stock_entry, dict):
-        work_order = stock_entry.get('work_order')
-    
-    # Batch dokümanını oluştur
-    batch_doc = frappe.get_doc(
-        {
-            "doctype": "Batch",
-            "batch_id": batch_id,
-            "item": row.item_code,
-            "reference_doctype": DOCTYPE_WORK_ORDER if work_order else stock_entry.doctype,
-            "reference_name": work_order if work_order else stock_entry.name,
-            "stock_entry_reference_doctype": stock_entry.doctype,
-            "stock_entry_reference_name": stock_entry.name,
-            "manufacturing_date": stock_entry.get(FIELD_POSTING_DATE),
-            "expiry_date": row.get("expiry_date"),
-            "stock_uom": row.get("stock_uom"),
-            "description": row.get("description"),
-        }
-    )
-
-    batch_doc.flags.ignore_permissions = True
-    batch_doc.insert()
-    return batch_doc.name
 
 
-def get_zebra_printer_for_user():
-    user = frappe.session.user
-
-    printer = frappe.db.get_value(
-        doctype=DOCTYPE_KTA_USER_ZEBRA_PRINTERS,
-        filters={"user": user, "disabled": 0},
-        fieldname="printer"
-    )
-
-    if printer is not None:
-        zebra_printer = frappe.get_doc(DOCTYPE_KTA_ZEBRA_PRINTERS, printer)
-        return zebra_printer
-    else:
-        frappe.msgprint("No default printer found for the current user.")
-        return None
 
 
-def format_kta_label_qty(qty):
-    if qty % 1 == 0:
-        return format_decimal(f"{qty:g}", locale="tr_TR")
-    else:
-        return format_decimal(f"{qty:.2f}", locale="tr_TR")
 
 
-def get_batch_from_stock_entry_detail(stock_entry_detail):
-    if not stock_entry_detail.get("serial_and_batch_bundle"):
-        return None
-
-    batch_no = frappe.db.get_value(
-        DOCTYPE_SERIAL_AND_BATCH_ENTRY,
-        filters={
-            FIELD_PARENT: stock_entry_detail.get("serial_and_batch_bundle"),
-            FIELD_PARENTTYPE: DOCTYPE_SERIAL_AND_BATCH_BUNDLE,
-            FIELD_PARENTFIELD: VALUE_ENTRIES,
-            FIELD_IS_OUTWARD: 0,
-            FIELD_WAREHOUSE: stock_entry_detail.get(FIELD_T_WAREHOUSE),
-            FIELD_BATCH_NO: ["is", "set"],
-            FIELD_DOCSTATUS: 1
-        },
-        fieldname=FIELD_BATCH_NO
-    )
-
-    if not batch_no:
-        frappe.throw(f"More than one batch found for Stock Entry Detail: {stock_entry_detail.name}")
-        return None
-
-    return batch_no
 
 
-def get_base_batch_for_work_order(work_order, item_code=None):
-    """
-    Work Order için base batch numarasını döndürür.
-    Eğer split edilmiş batch'ler varsa, onların base'ini kullanır.
-    """
-    if not work_order:
-        return None
-        
-    base_batch = get_base_batch_from_work_order(work_order, item_code)
-    if not base_batch:
-        return None
-    
-    # Base batch prefix'i bul (sonundaki 0001 gibi ardışık sayıları temizle)
-    # Ancak orijinal base batch'in kendisi rakamla bitiyorsa dikkatli olunmalı.
-    # Şimdilik basitçe full base batch'i kullanıyoruz, çünkü Work Order için yaratılan
-    # ilk batch "Base Batch" kabul ediliyor.
-    
-    return base_batch
 
 
-def get_base_batch_from_work_order(work_order, item_code=None):
-    if not work_order:
-        return None
-
-    filters = {"reference_doctype": DOCTYPE_WORK_ORDER, "reference_name": work_order}
-    if item_code:
-        filters["item"] = item_code
-
-    base_batches = frappe.get_all(
-        "Batch",
-        filters=filters,
-        pluck="name",
-        order_by="creation asc",
-        limit_page_length=1,
-    )
-
-    if base_batches:
-        return base_batches[0]
-
-    return None
 
 
-def split_manufacturing_batches(stock_entry):
-    """
-    Manufacturing Stock Entry için batch'leri paketleme miktarına göre böler.
-    Daha önce oluşturulmuş batch'lerin devamından sequence numarası verir.
-    
-    Örnek:
-    - İş emri base batch: 3506381
-    - İlk üretim: 150 adet -> 35063810001-35063810006 (6 paket)
-    - İkinci üretim: 50 adet -> 35063810007-35063810008 (2 paket)
-    """
-    if stock_entry.flags.in_split_process:
-        return
-    
-    stock_entry.flags.in_split_process = True
-
-    doc = stock_entry
-    if isinstance(stock_entry, str):
-        doc = frappe.get_doc(DOCTYPE_STOCK_ENTRY, stock_entry)
-
-    if not doc or doc.doctype != DOCTYPE_STOCK_ENTRY:
-        return
-
-    if doc.get("purpose") != "Manufacture":
-        return
-
-    packaging_cache = {}
-    for row in doc.get("items", []):
-        if not row.get(FIELD_IS_FINISHED_ITEM):
-            continue
-
-        bundle_name = row.get("serial_and_batch_bundle")
-        if not bundle_name:
-            continue
-
-        base_entry = _get_single_inward_batch_entry(bundle_name)
-        if not base_entry or not base_entry.get(FIELD_BATCH_NO):
-            continue
-
-        # Base batch numarasını al ve prefix'ini bul
-        # Örnek: 3506381 -> 3506381, 35063810001 -> 3506381
-
-        base_batch_prefix = get_base_batch_for_work_order(doc.work_order, row.item_code)
-        
-        # Eğer Work Order'dan base batch bulunamazsa, fallback olarak mevcut batch'i kullan
-        # Ama bu riskli, çünkü zaten split edilmiş bir batch olabilir.
-        if not base_batch_prefix:
-            base_batch = base_entry.get(FIELD_BATCH_NO)
-            base_batch_prefix = base_batch
-            
-            if not base_batch_prefix:
-                base_batch_prefix = base_entry.get(FIELD_BATCH_NO)
-
-
-        # Paketleme miktarını cache'den al veya DB'den oku
-        split_qty = packaging_cache.get(row.item_code)
-        if split_qty is None:
-            split_qty = _get_customer_packaging_qty(row.item_code)
-            packaging_cache[row.item_code] = split_qty
-
-        if not split_qty:
-            continue
-
-        # Yeni batch allocation'ları hazırla (devam eden sequence'den)
-        allocations = _prepare_batch_allocations(
-            row=row,
-            base_source_doc=doc,
-            base_batch_number=base_batch_prefix,
-            is_manufacturing=True
-        )
-
-        if not allocations:
-            continue
-
-        # Bundle'ı yeni batch'lerle güncelle
-        _update_bundle_safely(row, allocations)
 
 
 @frappe.whitelist()
