--- erpnextkta/erpnextkta/doctype/kta_depo_etiketleri/kta_depo_etiketleri.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2025, Framras AS and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("KTA Depo Etiketleri", {
-// 	refresh(frm) {
-
-// 	},
-// });
