--- erpnextkta/erpnextkta/doctype/kta_depo_etiketleri/test_kta_depo_etiketleri.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestKTADepoEtiketleri(FrappeTestCase):
-	pass
