--- erpnextkta/erpnextkta/doctype/kta_depo_etiketleri_bolme/kta_depo_etiketleri_bolme.json
@@ -1,39 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2025-07-08 14:07:36.878986",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "section_break_e01s",
-  "qty"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_e01s",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "qty",
-   "fieldtype": "Float",
-   "in_list_view": 1,
-   "label": "Qty",
-   "precision": "2"
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-07-08 14:30:44.481295",
- "modified_by": "Administrator",
- "module": "erpnextkta",
- "name": "KTA Depo Etiketleri Bolme",
- "owner": "Administrator",
- "permissions": [],
- "row_format": "Dynamic",
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
