--- erpnextkta/erpnextkta/doctype/kta_is_emri_etiketleri/kta_is_emri_etiketleri.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2025, Framras AS and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("KTA Is Emri Etiketleri", {
-// 	refresh(frm) {
-
-// 	},
-// });
