--- erpnextkta/erpnextkta/doctype/kta_is_emri_etiketleri/test_kta_is_emri_etiketleri.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestKTAIsEmriEtiketleri(FrappeTestCase):
-	pass
