--- erpnextkta/erpnextkta/doctype/kta_user_zebra_printers/kta_user_zebra_printers.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2025, Framras AS and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("KTA User Zebra Printers", {
-// 	refresh(frm) {
-
-// 	},
-// });
