--- erpnextkta/erpnextkta/doctype/kta_user_zebra_printers/kta_user_zebra_printers.json
@@ -1,65 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "field:user",
- "creation": "2025-04-02 18:41:42.742582",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "user",
-  "printer",
-  "disabled",
-  "section_break_bajg"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_bajg",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "user",
-   "fieldtype": "Link",
-   "label": "User",
-   "options": "User",
-   "unique": 1
-  },
-  {
-   "fieldname": "printer",
-   "fieldtype": "Link",
-   "label": "Zebra Printer",
-   "options": "KTA Zebra Printers"
-  },
-  {
-   "default": "0",
-   "fieldname": "disabled",
-   "fieldtype": "Check",
-   "label": "Disabled"
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2025-04-02 18:50:54.525450",
- "modified_by": "Administrator",
- "module": "erpnextkta",
- "name": "KTA User Zebra Printers",
- "naming_rule": "By fieldname",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
