--- erpnextkta/erpnextkta/doctype/kta_user_zebra_printers/test_kta_user_zebra_printers.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestKTAUserZebraPrinters(FrappeTestCase):
-	pass
