--- erpnextkta/erpnextkta/doctype/kta_zebra_printers/kta_zebra_printers.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class KTAZebraPrinters(Document):
-	pass
