--- erpnextkta/erpnextkta/doctype/kta_zebra_templates/kta_zebra_templates.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2025, Framras AS and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("KTA Zebra Templates", {
-// 	refresh(frm) {
-
-// 	},
-// });
