--- erpnextkta/erpnextkta/doctype/kta_zebra_templates/kta_zebra_templates.json
@@ -1,59 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "field:doctype_name",
- "creation": "2025-06-11 22:48:50.704524",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "section_break_ds1h",
-  "doctype_name",
-  "zebra_template"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_ds1h",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "doctype_name",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Doctype Name",
-   "reqd": 1,
-   "unique": 1
-  },
-  {
-   "fieldname": "zebra_template",
-   "fieldtype": "Text",
-   "label": "Zebra Template"
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2025-06-11 22:50:59.232077",
- "modified_by": "Administrator",
- "module": "erpnextkta",
- "name": "KTA Zebra Templates",
- "naming_rule": "By fieldname",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "row_format": "Dynamic",
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
