--- erpnextkta/erpnextkta/doctype/kta_zebra_templates/kta_zebra_templates.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class KTAZebraTemplates(Document):
-	pass
