--- erpnextkta/erpnextkta/doctype/kta_zebra_templates/test_kta_zebra_templates.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, Framras AS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestKTAZebraTemplates(FrappeTestCase):
-	pass
